Remove unused classes, methods, and files related to deprecated user and project management functionalities. Refactor imports and streamline DTO package structure.
